+++ source/objects/lander.lua
@@ -111,6 +111,8 @@ local function moveShip(lander, dt)
 		LANDER_VY = lander.vy
 		LANDER_VX = lander.vx
 	end
+	
+	-- lander.x = Cf.round(lander.x,0)
 end
 
 
@@ -290,7 +292,7 @@ end
 function Lander.create(name)
 	-- create a lander and return it to the calling sub
 	local lander = {}
-	lander.x = ORIGIN_X
+	lander.x = Cf.round(ORIGIN_X,0)
 	lander.y = GROUND[lander.x] - 8
 	lander.connectionID = nil	-- used by enet
 	-- 270 = up
@@ -330,6 +332,48 @@ end
 
 
 
+function Lander.reset(lander)
+-- resets a single lander. Used in multiplayer mode when you don't want to reset every lander.
+-- this function largely follows same behaviour as the CREATE function
+
+	lander.x = Cf.round(ORIGIN_X,0)
+	lander.y = GROUND[lander.x] - 8
+	-- lander.connectionID = nil	-- used by enet
+	-- 270 = up
+	lander.angle = 270
+	lander.vx = 0
+	lander.vy = 0
+	lander.engineOn = false
+	lander.leftEngineOn = false
+	lander.rightEngineOn = false
+	lander.onGround = false
+	-- Health in percent
+	lander.health = 100
+	lander.money = 0
+	lander.gameOver = false
+	lander.score = lander.x - ORIGIN_X
+	-- lander.name = name or CURRENT_PLAYER_NAME
+
+	-- mass
+	lander.mass = {}
+	-- base mass of lander
+	table.insert(lander.mass, 100)
+	-- volume in arbitrary units
+	lander.fuelCapacity = 25
+	-- start with a full tank
+	lander.fuel = lander.fuelCapacity
+	-- this is the mass of an empty tank
+	table.insert(lander.mass, 20)
+	-- this is the mass of the rangefinder (not yet purchased)
+	table.insert(lander.mass, 0)
+
+	-- modules
+	-- this will be strings/names of modules
+	lander.modules = {}
+end
+
+
+
 function Lander.getMass(lander)
 	-- return the mass of all the bits on the lander
     local result = 0
@@ -349,6 +393,7 @@ end
 
 function Lander.isOnLandingPad(lander, baseId)
 	-- returns a true / false value
+
     local baseDistance, _ = Fun.GetDistanceToClosestBase(lander.x, baseId)
     if baseDistance >= -80 and baseDistance <= 40 then
         return true
@@ -366,6 +411,21 @@ function Lander.hasUpgrade(lander, module)
 		end
 	end
 	return false
+end
+
+
+
+local function updateScore(lander)
+-- updates the lander score that is saved in the lander table
+-- this is the same as functions.CalculateScore(). Intention is to deprecate and remove that function and use this.
+-- this procedure does not return the score. It updates the lander table
+	
+	lander.score = lander.x - ORIGIN_X
+
+	if lander.score > GAME_SETTINGS.HighScore then
+		GAME_SETTINGS.HighScore = lander.score
+		Fun.SaveGameSettings() -- this needs to be refactored somehow, not save every change
+	end
 end
 
 
@@ -398,6 +458,7 @@ function Lander.update(lander, dt)
     moveShip(lander, dt)
     playSoundEffects(lander)
     checkForContact(lander, dt)
+	updateScore(lander)
 end
 
 
@@ -406,6 +467,9 @@ function Lander.draw()
 	-- draw the lander and flame
 	for landerId, lander in pairs(LANDERS) do
 		local sx, sy = 1.5, 1.5
+
+		assert(lander.x ~= nil)
+
 		local x = lander.x - WORLD_OFFSET
 		local y = lander.y
 		local ox = ship.width / 2